multiple improvement to import speed and export design
more simplification and abstraction of design to support many inputs.
This could have been a smaller more specific commit but things got out of hand.
This commit is being done at a functional point.  Using this code should produce mostly valid and functioning results both in the editor and in the game.