Preload table stuff, Alternative Bundle loading via Reference type, improved bundle name identification in loading processes
Calculate preload tables.  This is pretty much useless currently as it was developed to solve a load issue, however it doesn't solve that issue as proved by a test that Destroys a preload table on a functional bundle.

This update includes a bunch of stuff that makes the generated bundles follow more closely to normal bundle layouts however this is still not a fully functional result.
Material references are not honored for reasons that remain unknown.

Improve the identification of same named bundles currently loaded