Only create CAB & json map entries for root objects.
If an object like Transform is added as a sub-object, unity can
consider it orphaned and destroy it, which invalidates the object that
uses it.  That kind of problem may need to be fixed in the importer
eventually, but removing objects we didn't want to browse anyway reduces the problem.